rock-paper-scissors/requirements.txt
Flask>=2.0.2
datetime
azure.identity
azure.storage.blob